implement feedback in pull request

+++ skactiveml/pool/_cost_embedding_al.py
@@ -47,7 +47,7 @@ class CostEmbeddingAL(SingleAnnotatorPoolQueryStrategy):
         confident variant.
     missing_label: str or numeric, optional (default=MISSING_LABEL)
         Specifies the symbol that represents a missing label.
-    random_state: numeric or np.random.RandomState, optional
+    random_state : int or np.random.RandomState, optional
     (default=None)
         Random state for annotator selection.
     embed_dim : int, optional (default=None)
@@ -77,7 +77,9 @@ class CostEmbeddingAL(SingleAnnotatorPoolQueryStrategy):
         missing_label=MISSING_LABEL,
         random_state=None,
     ):
-        super().__init__(missing_label=missing_label, random_state=random_state)
+        super().__init__(
+            missing_label=missing_label, random_state=random_state
+        )
         self.classes = classes
         self.base_regressor = base_regressor
         self.cost_matrix = cost_matrix
@@ -226,11 +228,11 @@ def _alce(
         Array of class labels.
     embed_dim: int
         Dimension of the embedding.
-    sample_weight: array-like, shape (n_samples)
+    sample_weight : array-like, shape (n_samples)
         Weights for uncertain annotators.
-    missing_label: scalar | string | np.nan | None
+    missing_label : scalar | string | np.nan | None
         Value to represent a missing label.
-    random_state: numeric | np.random.RandomState
+    random_state : int | np.random.RandomState
         Random state for annotator selection.
     mds_params : dict
         For further information, see
@@ -455,7 +457,9 @@ def _smacof_single_p(
             disparities[similarities == 0] = 0
 
         # Compute stress
-        _stress = (W.ravel() * ((dis.ravel() - disparities.ravel()) ** 2)).sum()
+        _stress = (
+            W.ravel() * ((dis.ravel() - disparities.ravel()) ** 2)
+        ).sum()
         _stress /= 2
 
         # Update X using the Guttman transform
