+++ skactiveml/pool/_epistemic_uncertainty_sampling.py
@@ -39,7 +39,7 @@ class EpistemicUncertaintySampling(SingleAnnotatorPoolQueryStrategy):
     missing_label : scalar or string or np.nan or None, optional
     (default=MISSING_LABEL)
         Value to represent a missing label.
-    random_state : numeric or np.random.RandomState
+    random_state : int or np.random.RandomState
         The random state to use.
 
     References
@@ -52,7 +52,9 @@ class EpistemicUncertaintySampling(SingleAnnotatorPoolQueryStrategy):
     def __init__(
         self, precompute=False, missing_label=MISSING_LABEL, random_state=None
     ):
-        super().__init__(missing_label=missing_label, random_state=random_state)
+        super().__init__(
+            missing_label=missing_label, random_state=random_state
+        )
         self.precompute = precompute
 
     def query(
