+++ skactiveml/pool/_quire.py
@@ -32,7 +32,7 @@ class Quire(SingleAnnotatorPoolQueryStrategy):
         Any further parameters are passed directly to the metric function.
     missing_label : scalar or string or np.nan or None, default=MISSING_LABEL
         Value to represent a missing label.
-    random_state : numeric or np.random.RandomState, optional (default=None)
+    random_state : int or np.random.RandomState, optional (default=None)
         The random state to use.
 
     References
@@ -53,7 +53,9 @@ class Quire(SingleAnnotatorPoolQueryStrategy):
         missing_label=MISSING_LABEL,
         random_state=None,
     ):
-        super().__init__(missing_label=missing_label, random_state=random_state)
+        super().__init__(
+            missing_label=missing_label, random_state=random_state
+        )
         self.classes = classes
         self.lmbda = lmbda
         self.metric = metric
@@ -109,7 +111,13 @@ class Quire(SingleAnnotatorPoolQueryStrategy):
         """
         # --- Validation -----------------------------------------------------
         # Check standard parameters.
-        (X, y, candidates, batch_size, return_utilities,) = self._validate_data(
+        (
+            X,
+            y,
+            candidates,
+            batch_size,
+            return_utilities,
+        ) = self._validate_data(
             X=X,
             y=y,
             candidates=candidates,
