+++ skactiveml/pool/_uncertainty_sampling.py
@@ -38,7 +38,7 @@ class UncertaintySampling(SingleAnnotatorPoolQueryStrategy):
         `least_confident` and `margin_sampling` variant.
     missing_label : scalar or string or np.nan or None, default=np.nan
         Value to represent a missing label.
-    random_state : numeric or np.random.RandomState
+    random_state : int or np.random.RandomState
         The random state to use.
 
     References
@@ -61,7 +61,9 @@ class UncertaintySampling(SingleAnnotatorPoolQueryStrategy):
         missing_label=MISSING_LABEL,
         random_state=None,
     ):
-        super().__init__(missing_label=missing_label, random_state=random_state)
+        super().__init__(
+            missing_label=missing_label, random_state=random_state
+        )
         self.method = method
         self.cost_matrix = cost_matrix
